LC-find-x-sum-of-all-k-long-subarrays-ii [Runtime: N/A; Memory: N/A]

+++ heap-priority-queue/find-x-sum-of-all-k-long-subarrays-ii.cpp
@@ -4,9 +4,9 @@ public:
         vector<long long>ans;
         int n=nums.size();
         for(int i=0; i<n-k+1; i++){
-            unordered_map<int,int>freq;
-            priority_queue<pair<int,int>>pq;
-            int sum=0;
+            unordered_map<long long,long long>freq;
+            priority_queue<pair<long long,long long>>pq;
+            long long sum=0;
             int j=0;
             vector<int>temp(nums.begin()+i, nums.begin()+k+i);
             for(auto t:temp)freq[t]++;
